spdx: Move PackageFromDirectory to implementation
This commit moves the spdx.PackageFromDirectory() to the implementation
to make it mockable.
